Working version of image simulation code
Plus some initial image files and a notebook to play around with the results
